api/requirements.txt
fastapi==0.115.8
numpy
python-multipart
opencv_python==4.11.0.86
Pillow==11.1.0
uvicorn==0.34.0
websocket_client==1.8.0
